build rpcs.json
