+++ pages/index.js
@@ -3,18 +3,10 @@ import Head from "next/head";
 import { useRouter } from "next/router";
 import Layout from "../components/Layout";
 import Chain from "../components/chain";
-import { fetcher, populateChain } from "../utils";
+import { generateChainData } from "../utils/fetch";
 
 export async function getStaticProps() {
-  const chains = await fetcher("https://chainid.network/chains.json");
-  const chainTvls = await fetcher("https://api.llama.fi/chains");
-
-  const sortedChains = chains
-    .filter((c) => c.name !== "420coin") // same chainId as ronin
-    .map((chain) => populateChain(chain, chainTvls))
-    .sort((a, b) => {
-      return (b.tvl ?? 0) - (a.tvl ?? 0);
-    });
+  const sortedChains = await generateChainData()
 
   return {
     props: {
