+++ pages/zh/index.js
@@ -3,7 +3,7 @@ import Head from "next/head";
 import { useRouter } from "next/router";
 import Layout from "../../components/Layout";
 import Chain from "../../components/chain";
-import { fetcher, populateChain } from "../../utils";
+import { fetcher, populateChain } from "../../utils/fetch";
 
 export async function getStaticProps() {
   const chains = await fetcher("https://chainid.network/chains.json");
